Scaling EXP with monster lvl
Monsters give fixed exp based on their lvl, 20% increase over the previous lvl each time. lvl 1 starts at 10 exp. Exp tnl adjusted to match the same number of EM kills for a lvl 75.

Known Issue: when exp to lvl up gets above 65,537 it overflows and resets exp progress for that lvl. I think I missed some uint16s that need to be uing32s to get it to work.

+++ src/map/packets/char_stats.cpp
@@ -43,8 +43,8 @@ CCharStatsPacket::CCharStatsPacket(CCharEntity* PChar)
     ref<uint8>(0x0E) = PChar->GetSJob();
     ref<uint8>(0x0F) = PChar->GetSLevel();
 
-    ref<uint16>(0x10) = PChar->jobs.exp[PChar->GetMJob()];
-    ref<uint16>(0x12) = charutils::GetExpNEXTLevel(PChar->jobs.job[PChar->GetMJob()]);
+    ref<uint32>(0x10) = PChar->jobs.exp[PChar->GetMJob()];
+    ref<uint32>(0x12) = charutils::GetExpNEXTLevel(PChar->jobs.job[PChar->GetMJob()]);
 
     memcpy(data + (0x14), &PChar->stats, 14); // TODO: с merits это не прокатит
